requirements.txt
aiogram==2.25.1 
requests==2.31.0 
sqlite3  
uuid  
asyncio  
threading
python-dotenv==1.0.0